Refactor: testUtil 파일 분리
